+++ core/layout/meta_footer.php
@@ -326,6 +326,7 @@ if(isset($r['due_ti'])){?>
   document.addEventListener("DOMContentLoaded",function(event){
 <?php if($args[0]=='edit'||$args[0]=='compose'||$args[0]=='reply'||($view=='accounts'||$view=='orders'||$view=='bookings'||$view=='newsletters'||$view=='messages'&&$args[0]=='settings')){?>
     $('.summernote').summernote({
+      isNotSplitEdgePoint:true,
       height:300,
       tabsize:2,
       popover:{