+++ index.php
@@ -7,12 +7,14 @@
  * @author     Dennis Suitters <dennis@diemen.design>
  * @copyright  2014-2019 Diemen Design
  * @license    http://opensource.org/licenses/MIT  MIT License
- * @version    0.0.1
+ * @version    0.0.13
  * @link       https://github.com/DiemenDesign/AuroraCMS
  * @notes      This PHP Script is designed to be executed using PHP 7+
+ * @changes    v0.0.13 Add X-Clacks Silent Attribute Header, Add you site to the list: https://xclacksoverhead.org/dearheart/review
  */
 mb_internal_encoding("UTF-8");
 mb_http_output("UTF-8");
+header("X-Clacks-Overhead: GNU Terry Pratchett"); // The unseen, silent tribute to those we have lost.
 ini_set('session.use_trans_sid',0);
 ini_set('session.use_cookies',1);
 ini_set('session.use_only_cookies',1);